Initial Commit
Create manifest and beginnings of interface to boot and mute users.
Uses async/await and a wait function to find DOM elements and work
on them.

For mute it loops through users and tries to turn off each. On boot
it just boots one by one until room count is 1.